chore: refactor the IsolatedTestDirWithAutoCleanup and break import cycle
