+++ op-deployer/pkg/deployer/artifacts/download_test.go
@@ -8,15 +8,14 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"path"
 	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/minio/sha256-simd"
 
+	"github.com/ethereum-optimism/optimism/op-service/testutils"
 	"github.com/ethereum/go-ethereum/common"
-
 	"github.com/stretchr/testify/require"
 )
 
@@ -45,7 +44,7 @@ func TestDownloadArtifacts_MockArtifacts(t *testing.T) {
 		URL: artifactsURL,
 	}
 
-	testCacheDir := IsolatedTestDirWithAutoCleanup(t)
+	testCacheDir := testutils.IsolatedTestDirWithAutoCleanup(t)
 
 	t.Run("success", func(t *testing.T) {
 		fs, err := Download(ctx, loc, nil, testCacheDir)
@@ -113,27 +112,12 @@ func TestDownloadArtifacts_MockArtifacts(t *testing.T) {
 	})
 }
 
-// duplicate of testutil.IsolatedTestDirWithAutoCleanup to break the import cycle
-// TODO: shift them to a shared package
-func IsolatedTestDirWithAutoCleanup(t *testing.T) string {
-	basePath := os.Getenv("OP_DEPLOYER_TEST_DIR")
-	if basePath == "" {
-		basePath = "./.tests"
-	}
-	dir := path.Join(basePath, t.Name())
-	require.NoError(t, os.MkdirAll(dir, 0755))
-	t.Cleanup(func() {
-		require.NoError(t, os.RemoveAll(dir))
-	})
-	return dir
-}
-
 func TestDownloadArtifacts_TaggedVersions(t *testing.T) {
 	tags := []string{
 		"op-contracts/v1.6.0",
 		"op-contracts/v1.7.0-beta.1+l2-contracts",
 	}
-	testCacheDir := IsolatedTestDirWithAutoCleanup(t)
+	testCacheDir := testutils.IsolatedTestDirWithAutoCleanup(t)
 	for _, tag := range tags {
 		t.Run(tag, func(t *testing.T) {
 			t.Parallel()
